Update supported-platforms.mdx
"Depending" rather than "dependent" as you "may" need to provide.
"We support X and later" in line with commonly-distributed technical documentation. "And" rather than "or" as we support all versions from X to later. "Later" rather than "above" to clarify that time rather than version number is the key determiner.